+++ packages/example-express/.eslintrc.js
@@ -1,0 +1,15 @@
+module.exports = {
+  parser: '@typescript-eslint/parser',
+  extends: [
+    'eslint:recommended',
+    'plugin:@typescript-eslint/recommended'
+  ],
+  plugins: ['@typescript-eslint'],
+  env: {
+    node: true,
+    es6: true
+  },
+  rules: {
+    '@typescript-eslint/no-explicit-any': 'off'
+  }
+};